Reorganize file structure # 26
